+++ iotedgedev/__init__.py
@@ -1,8 +1,8 @@
-# -*- coding: utf-8 -*-
-
 """Top-level package for Azure IoT Edge Dev Tool."""
+
+from . import version_catch
 
 __author__ = 'Microsoft Corporation'
 __email__ = 'vsciet@microsoft.com'
-__version__ = '2.1.6'
+__version__ = '3.0.0'
 __AIkey__ = '95b20d64-f54f-4de3-8ad5-165a75a6c6fe'