+++ iotedgedev/version_catch.py
@@ -1,0 +1,3 @@
+from .version import PY3
+
+assert PY3, "Python 2 is no longer supported on this project. Please upgrade to Python 3.6 or higher."